requirements.txt
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
python-telegram-bot==20.6
SQLAlchemy==1.4.27
asyncio==4.0.0
requests==2.26.0
pytest==6.2.5
gunicorn==20.1.0
python-dotenv==0.19.2
Werkzeug==2.0.3
bcrypt==4.0.1
Pillow==10.0.1
qrcode[pil]==7.4.2
psycopg2-binary==2.9.7